refactor: move github bot in a separate package
