+++ package.json
@@ -14,15 +14,15 @@
 	},
 	"license": "MIT",
 	"dependencies": {
-		"@types/node": "^18.15.11",
-		"msw": "^1.2.1",
-		"tsx": "^3.12.6",
-		"typescript": "^4.9.5"
+		"@types/node": "~18.15.11",
+		"tsx": "~3.12.6",
+		"typescript": "~4.9.5"
 	},
 	"devDependencies": {
 		"@martin-kolarik/eslint-config": "^7.3.2",
 		"@typescript-eslint/eslint-plugin": "^8.11.0",
 		"eslint": "^8.57.1",
 		"vitest": "^2.1.2"
-	}
+	},
+	"packageManager": "pnpm@9.15.0+sha512.76e2379760a4328ec4415815bcd6628dee727af3779aaa4c914e3944156c4299921a89f976381ee107d41f12cfa4b66681ca9c718f0668fa0831ed4c6d8ba56c"
 }
