+++ packages/bot-utils/package.json
@@ -18,6 +18,7 @@
 	},
 	"devDependencies": {
 		"@types/yargs-parser": "^21.0.0",
+		"msw": "^2.7.0",
 		"pkgroll": "^1.9.0"
 	}
 }
