+++ packages/bot-utils/src/index.ts
@@ -6,3 +6,5 @@ export * from './measurements-request.js';
 export * from './post.js';
 export * from './types.js';
 export * from './utils.js';
+export * from './help.js';
+export * from './response.js';
