+++ packages/github/.env.sample
@@ -1,0 +1,8 @@
+# A secret API Key that will allow authenticating http requests to the endpoint
+GITHUB_BOT_API_KEY=
+# The github username of the bot (globalping in this case for the prod version)
+GITHUB_BOT_HANDLE=
+# A Github Personal Access Token for the Github bot user account, with the public_repo scope selected
+GITHUB_PERSONAL_ACCESS_TOKEN=
+
+GLOBALPING_TOKEN=
