+++ packages/github/package.json
@@ -1,0 +1,20 @@
+{
+	"name": "github-bot",
+	"private": true,
+	"version": "1.0.0",
+	"main": "./src/index.ts",
+	"scripts": {
+		"test": "echo \"Error: no test specified\" && exit 1"
+	},
+	"dependencies": {
+		"@globalping/bot-utils": "workspace:*",
+		"octokit": "^3.0.0",
+		"raw-body": "^2.5.2",
+		"remark": "^14.0.3",
+		"strip-markdown": "^5.0.1",
+		"uuid": "^9.0.0"
+	},
+	"devDependencies": {
+		"@types/uuid": "^9.0.2"
+	}
+}
