+++ packages/github/src/index.ts
@@ -1,0 +1,2 @@
+export * from './bot.js';
+export * from './types.js';
