+++ packages/github/tsconfig.json
@@ -1,0 +1,46 @@
+{
+	"include": [
+		"**/*.ts",
+		"**/*.json"
+	],
+	"exclude": [
+		"node_modules"
+	],
+	"compilerOptions": {
+
+		/* Projects */
+		// "incremental": true,
+
+		/* Language and Environment */
+		"target": "esnext",
+		"useDefineForClassFields": true,
+		"composite": true,
+
+		/* Modules */
+		"module": "esnext",
+		"moduleResolution": "node",
+		"resolveJsonModule": true,
+
+		/* Emit */
+		"outDir": "./types",
+
+		/* Interop Constraints */
+		"isolatedModules": true,
+		"allowSyntheticDefaultImports": true,
+		"esModuleInterop": true,
+		"forceConsistentCasingInFileNames": true,
+
+		/* Type Checking */
+		"strict": true,
+		// "types": ["vitest/importMeta"],
+		// "isolatedModules": true, // Uncomment if you're running into an esbuild transform error
+		"noUnusedLocals": true,
+		"noUnusedParameters": true,
+		"noImplicitReturns": true,
+		"noFallthroughCasesInSwitch": true,
+
+		/* Completeness */
+		"skipLibCheck": true // Skip type checking all .d.ts files.
+	},
+	"$schema": "https://json.schemastore.org/tsconfig"
+}
