+++ packages/slack/package.json
@@ -15,18 +15,13 @@
 		"@globalping/bot-utils": "workspace:*",
 		"@slack/bolt": "^3.22.0",
 		"discord-bot": "workspace:*",
+		"github-bot": "workspace:*",
 		"dotenv": "^16.0.3",
 		"knex": "^2.5.1",
-		"mysql": "^2.18.1",
-		"octokit": "^3.0.0",
-		"raw-body": "^2.5.2",
-		"remark": "^14.0.3",
-		"strip-markdown": "^5.0.1",
-		"uuid": "^9.0.0"
+		"mysql": "^2.18.1"
 	},
 	"devDependencies": {
 		"@slack/web-api": "^6.13.0",
-		"@types/uuid": "^9.0.2",
 		"pino-pretty": "^10.0.0"
 	}
 }
