--- packages/slack/src/github/common.ts
@@ -1,5 +1,0 @@
-import { config } from '../config.js';
-
-export function githubHandle (): string {
-	return config.githubBotHandle;
-}
