+++ packages/slack/src/routes.ts
@@ -1,13 +1,15 @@
 import { CustomRoute } from '@slack/bolt';
-import { ParamsIncomingMessage } from '@slack/bolt/dist/receivers/ParamsIncomingMessage.js';
+import { initBot } from 'github-bot';
 import * as fs from 'node:fs';
-import { IncomingMessage, ServerResponse } from 'node:http';
 import path from 'node:path';
 import { fileURLToPath } from 'node:url';
 
 import { CALLBACK_PATH, oauth } from './auth.js';
+import { config } from './config.js';
 import { knex } from './db.js';
-import { githubHandler } from './github/handler.js';
+import { logger } from './utils.js';
+
+const githubBot = initBot(config, logger);
 
 export const routes: CustomRoute[] = [
 	{
@@ -89,14 +91,11 @@ export const routes: CustomRoute[] = [
 	{
 		path: '/github-bot',
 		method: [ 'POST' ],
-		handler: githubHandler,
+		handler: (req, res) => githubBot.HandleRequest(req, res),
 	},
 	{
 		path: CALLBACK_PATH,
 		method: [ 'GET' ],
-		handler: (
-			req: ParamsIncomingMessage,
-			res: ServerResponse<IncomingMessage>,
-		) => oauth.OnCallback(req, res),
+		handler: (req, res) => oauth.OnCallback(req, res),
 	},
 ];
